Keep a register's bits as bits in the instruction writer
The register-reuse wall, from the register side. The instruction writer
declares each register once, and fxc reuses them freely: a loop counter,
a comparison mask and a float can share one component in turn. Declared
int, the float truncated; declared float, the mask's bits were read as
a number.

So a component now has a storage model, not the register a type. A
register is declared int only when nothing float ever touches any
component it writes. Otherwise it is float, and a component in it
holds either its value - an integer as the equal float, converted on
the way through itof and ftoi as before - or, once a comparison or a
bitwise operator has had it, its bits, which the integer instructions
read with asint and write with asfloat. Reinterpreting rather than
converting is what keeps a mask a mask. A moved immediate is typed by
the instructions that go on to read the component, the way the AST
writer does it, and a mov between storages converts.

Only the components that were wrong change output: two registers flip
from int to float, the mask arithmetic in step_mask gains asint and
asfloat, and shadow_pcf's -1.0f is -1 again. Every shader now computes
the same as its original from both writers - KnownDifferences is empty
for the first time - and the full suite runs with nothing skipped:
1265 passed, 0 failed, 0 skipped.

+++ HlslDecompiler.Tests/ShaderSources/ps_4_0_instruction/get_dimensions.fx
@@ -12,7 +12,7 @@ float4 main(float4 texcoord : TEXCOORD) : SV_Target
 	tex.GetDimensions(0, dimensions0.x, dimensions0.y, dimensions0.w);
 	r1 = dimensions0;
 	r0.xy = r0.xy % r1.xy;
-	r0.zw = float2(0, 0);
+	r0.zw = int2(0, 0);
 	o = tex.Load(r0.xyz);
 
 	return o;
